+++ CultureCenter/src/main/java/com/culture/demo/controller/MemberController.java
@@ -4,6 +4,7 @@ import java.util.Map;
 
 import javax.servlet.http.HttpServletResponse;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
@@ -22,6 +23,9 @@ import lombok.extern.log4j.Log4j;
 @Log4j
 @AllArgsConstructor
 public class MemberController {
+	
+	@Autowired
+	private MemberService service;
 
 	private PasswordEncoder passwordEncoder;
 
@@ -140,6 +144,58 @@ public class MemberController {
 		}
         return "redirect:../login/index.do";
 	}
+	
+	@GetMapping("/login/idCheck")
+	@ResponseBody
+	public String idcheck(String memberId) {
+		String chk = " ";
+		int result = 0;
+		log.info(">memberID = " + memberId);
+	    result = service.idCheck(memberId);
+	    log.info(">memberID 중복여부 = " + (result >= 1 ? "중복" : "중복아님"));
+	    if (result >= 1) {
+	        chk = "redundancy"; // 중복
+	    } else if (result == 0) {
+	        chk = "noredundancy"; // 중복아님
+	    }
+	    return chk;
+	}
+
+	// 아이디 찾기
+	@PostMapping("/login/findId.do")
+	@ResponseBody
+	public String findid(String name, String phone) {
+		log.info(">findId =" + name +" / "+ phone);
+
+		String foundId = service.findId(name, phone);
+		log.info(">findIdfindId = " + foundId);
+
+		return foundId;
+	}
+
+	// 비밀번호 찾기
+	@PostMapping("/login/findPw.do")
+	@ResponseBody
+	public String findPw(String id, String phone) {
+		log.info(">findId =" + id + " / " + phone);
+
+		String foundPw = service.findPW(id, phone);
+		log.info(">findIdfindPw = " + foundPw);
+
+		return foundPw;
+	}
+	// 아이디 찾기 이동
+		@GetMapping("/login/findId.do")
+		public String findId() throws Exception {
+			log.info("> LoginController findId.do......");
+			return "login/find_id";
+		}
+		// 비밀번호 찾기 이동
+		@GetMapping("/login/findPw.do")
+		public String findPw() throws Exception {
+			log.info("> LoginController findPw.do......");
+			return "login/find_pw";
+		}
 	
 	
 }
